+++ Source/UnrealCAPI/Public/uec_api.h
@@ -22,7 +22,7 @@
 #endif
 
 #define UEC_ABI_MAJOR 1u
-#define UEC_ABI_MINOR 16u
+#define UEC_ABI_MINOR 17u
 
 #ifdef __cplusplus
 extern "C" {
@@ -66,7 +66,8 @@ enum {
     UEC_CAPABILITY_ASSETS = UINT64_C(1) << 9,
     UEC_CAPABILITY_ASYNC_ASSETS = UINT64_C(1) << 10,
     UEC_CAPABILITY_LEVEL_TRAVEL = UINT64_C(1) << 11,
-    UEC_CAPABILITY_PLAYER_FLOW = UINT64_C(1) << 12
+    UEC_CAPABILITY_PLAYER_FLOW = UINT64_C(1) << 12,
+    UEC_CAPABILITY_INPUT = UINT64_C(1) << 13
 };
 
 typedef struct uec_context uec_context;
@@ -189,6 +190,12 @@ typedef struct uec_api {
                                         uec_actor* pawn);
     uec_result (UEC_CALL *set_controller_view_target)(uec_actor* controller,
                                                       uec_actor* view_target);
+    uec_result (UEC_CALL *get_input_key_down)(uec_actor* controller,
+                                              uec_string_view key_name,
+                                              uec_bool* out_down);
+    uec_result (UEC_CALL *get_input_key_value)(uec_actor* controller,
+                                               uec_string_view key_name,
+                                               double* out_value);
     uec_result (UEC_CALL *get_default_world)(uec_context* context, uec_world** out_world);
     uec_result (UEC_CALL *release_world)(uec_world* world);
     uec_result (UEC_CALL *spawn_actor)(uec_world* world,